Remove postselection from the workflow
Currently, the post-selection has no effect because it runs after
configuration recovery, which itself ensures that each bitstring matches
the Hamming weight that is required.  In general, post-selection is only
necessary during the first iteration of SQD, before occupancy data is known.
However, in this demo, we seed with known occupancy data, so
configuration recovery can be performed immediately, and post-selection is
not needed.

This PR thus removes post-selection from the main loop.